requirements.txt
accelerate==1.13.0
annotated-doc==0.0.4
anyio==4.13.0
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.2
cuda-bindings==13.2.0
cuda-pathfinder==1.5.2
cuda-toolkit==13.0.2
diffusers==0.37.1
exceptiongroup==1.3.1
filelock==3.25.2
fsspec==2026.3.0
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.10.1
idna==3.11
importlib_metadata==9.0.0
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
nvidia-cublas==13.1.0.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti==13.0.85
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc==13.0.88
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime==13.0.96
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cudnn-cu13==9.19.0.56
nvidia-cufft==12.0.0.61
nvidia-cufft-cu12==11.0.2.54
nvidia-cufile==1.15.1.6
nvidia-curand==10.4.0.35
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver==12.0.4.66
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse==12.6.3.3
nvidia-cusparse-cu12==12.1.0.106
nvidia-cusparselt-cu13==0.8.0
nvidia-nccl-cu12==2.21.5
nvidia-nccl-cu13==2.28.9
nvidia-nvjitlink==13.0.88
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvshmem-cu13==3.4.5
nvidia-nvtx==13.0.85
nvidia-nvtx-cu12==12.1.105
packaging==26.0
peft==0.18.1
pillow==12.2.0
psutil==7.2.2
Pygments==2.20.0
PyYAML==6.0.3
regex==2026.4.4
requests==2.33.1
rich==15.0.0
safetensors==0.7.0
shellingham==1.5.4
sympy==1.13.1
tokenizers==0.22.2
torch==2.5.1+cu121
torchvision==0.20.1+cu121
tqdm==4.67.3
transformers==5.5.4
triton==3.1.0
typer==0.24.1
typing_extensions==4.15.0
urllib3==2.6.3
zipp==3.23.1
